Fix 8 bugs from codebase review (Wave 1)
- Wire protocol Written size u32 → u64 to prevent truncation on
  copy_file_range/remap_file_range returns exceeding 4GB
- Loopback IP exhaustion now returns error instead of silently
  reusing 127.0.0.2 (would cause IP conflicts)
- Remove security.capability xattr fast-path that returned ENODATA
  for all files, hiding real capabilities
- Check e2fsck exit code before resize2fs (exit >= 4 means
  uncorrectable filesystem errors)
- slirp4netns stdout/stderr changed from Stdio::piped() to
  Stdio::null() to prevent pipe buffer deadlock
- Check truncate exit code in create_disk_from_dir
- parse_size uses checked_mul to prevent silent overflow
- Delete dead code mount_vsock_with_readers in fc-agent

Tested: cargo test -p fuse-pipe --lib (42 pass), cargo test -p fcvm --lib (48 pass)

+++ fuse-pipe/src/protocol/response.rs
@@ -31,7 +31,7 @@ pub enum VolumeResponse {
     Data { data: Vec<u8> },
 
     /// Number of bytes written.
-    Written { size: u32 },
+    Written { size: u64 },
 
     /// File opened response.
     Opened { fh: u64, flags: u32 },
